renamed metrics_client_ca.go to metrics.go
